microservices/traffic_monitoring/requirements.txt
flask
flask-cors
kafka-python
requests
gunicorn>=20.1.0
cryptography
spyne
lxml
